+++ posts-app/src/widgets/CommentList/ui/CommentList.tsx
@@ -14,11 +14,9 @@ type CommentListProps = {
 }
 
 function CommentList({ comments, open, toogle }: CommentListProps) {
-	const handleToogle = () => toogle()
-
 	return (
 		<div>
-			<div className={s.accordionSummary} onClick={handleToogle}>
+			<div className={s.accordionSummary} onClick={toogle}>
 				<Typography fontWeight="bold">Комментарии</Typography>
 				<ExpandMoreIcon />
 			</div>
