+++ posts-app/src/widgets/LayoutHeader/LayoutHeader.tsx
@@ -3,7 +3,7 @@ import { useState } from "react"
 import { NavLink } from "react-router"
 import ThemeSwitcher from "../../features/ThemeSwitcher/ui/ThemeSwitcher"
 import ButtonToogle from "../../shared/ui/Button/ButtonToogle.tsx"
-import { Modal } from "../../shared/ui/Modal/Modal.tsx"
+import Modal from "../../shared/ui/Modal/Modal.tsx"
 import Navigation from "../../shared/ui/Navigation/Navigation.tsx"
 import s from "./LayoutHeader.module.css"
